+++ Cargo.toml
@@ -7,7 +7,6 @@ members = [
    "examples/hello-world/hello-indexer",
    "examples/hello-world/hello-world-data",
    "examples/hello-world/hello-world-node",
-   "packages/fuel-indexer",
    "packages/fuel-indexer-api-server",
    "packages/fuel-indexer-benchmarks",
    "packages/fuel-indexer-database",
@@ -22,10 +21,12 @@ members = [
    "packages/fuel-indexer-schema",
    "packages/fuel-indexer-tests",
    "packages/fuel-indexer-tests/components/fuel-node",
-   "packages/fuel-indexer-tests/indexers/fuel-indexer-test",
    "packages/fuel-indexer-tests/components/web-api",
+   "packages/fuel-indexer-tests/indexers/fuel-indexer-test",
+   "packages/fuel-indexer-tests/indexers/simple-wasm/simple-wasm",
    "packages/fuel-indexer-types",
    "packages/fuel-indexer-utils",
+   "packages/fuel-indexer",
    "plugins/forc-index",
    "plugins/forc-postgres",
 ]
