+++ packages/fuel-indexer-lib/Cargo.toml
@@ -14,7 +14,7 @@ anyhow = "1.0"
 async-graphql-parser = { workspace = true }
 bincode = { workspace = true }
 clap = { features = ["cargo", "derive", "env"], workspace = true }
-fuels = { workspace = true }
+fuel-indexer-types = { workspace = true }
 http = { version = "0.2", default-features = false }
 lazy_static = { version = "1.4" }
 proc-macro2 = "1.0"
