+++ packages/fuel-indexer-macros/src/lib.rs
@@ -21,8 +21,7 @@ pub fn indexer(attrs: TokenStream, item: TokenStream) -> TokenStream {
 
 #[cfg(test)]
 mod tests {
-    #[allow(unused)]
-    #[ignore = "Currently fails due to dependency issues in fuel-indexer-macros"]
+    #[test]
     fn test_success_and_failure_macros() {
         let t = trybuild::TestCases::new();
         let manifest_dir = env!("CARGO_MANIFEST_DIR");
