+++ packages/fuel-indexer-tests/indexers/simple-wasm/bad_simple_wasm.yaml
@@ -1,0 +1,7 @@
+namespace: test_namespace
+identifier: test_simple_wasm
+abi: packages/fuel-indexer-tests/contracts/simple-wasm/out/debug/contracts-abi.json
+graphql_schema: schema.graphql
+contract_id: ~
+module:
+  wasm: simple_wasm.wasm
