+++ packages/fuel-indexer-tests/indexers/simple-wasm/simple-wasm/.cargo/config
@@ -1,0 +1,2 @@
+[build]
+target = "wasm32-unknown-unknown"
