+++ packages/fuel-indexer-tests/indexers/simple-wasm/simple-wasm/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "simple-wasm"
+version = "0.0.0"
+edition = "2021"
+publish = false
+
+[lib]
+crate-type = ['cdylib']
+
+[dependencies]
+fuel-indexer-utils = { workspace = true }
+fuel-tx = { workspace = true }
+fuels = { workspace = true }
+serde = { workspace = true }
+sha2 = { version = "0.10" }
