+++ packages/fuel-indexer-tests/indexers/simple-wasm/simple_wasm.yaml
@@ -1,0 +1,7 @@
+namespace: test_namespace
+identifier: simple_wasm_executor
+abi: packages/fuel-indexer-tests/contracts/simple-wasm/out/debug/contracts-abi.json
+graphql_schema: packages/fuel-indexer-tests/indexers/simple-wasm/schema/schema.graphql
+contract_id: ~
+module:
+  wasm: target/wasm32-unknown-unknown/release/simple_wasm.wasm
