+++ packages/fuel-indexer-tests/tests/lib.rs
@@ -1,10 +1,10 @@
 #[cfg(all(feature = "postgres"))]
-mod graphql_api_postgres;
+mod graphql_server;
 
 #[cfg(all(feature = "postgres"))]
-mod indexing_postgres;
+mod indexing;
 
 mod service;
 
 #[cfg(all(feature = "postgres"))]
-mod web_api_postgres;
+mod web_server;
