+++ packages/fuel-indexer-tests/trybuild/fail_if_attribute_abi_arg_includes_invalid_type.rs
@@ -15,7 +15,7 @@ mod indexer {
 }
 
 fn main() {
-    use fuels::core::abi_encoder::ABIEncoder;
+    use fuels::core::codec::ABIEncoder;
     use fuels::types::traits::Tokenizable;
 
     let s = SomeEvent {
