+++ packages/fuel-indexer-tests/trybuild/pass_if_indexer_is_valid_multi_type.rs
@@ -7,6 +7,9 @@ fn ff_log_data(_inp: ()) {}
 #[no_mangle]
 fn ff_put_object(_inp: ()) {}
 
+#[no_mangle]
+fn ff_put_many_to_many_record(_inp: ()) {}
+
 #[indexer(manifest = "packages/fuel-indexer-tests/trybuild/simple_wasm.yaml")]
 mod indexer {
     fn function_one(event: SomeEvent) {
@@ -26,7 +29,7 @@ mod indexer {
 }
 
 fn main() {
-    use fuels::core::abi_encoder::ABIEncoder;
+    use fuels::core::codec::ABIEncoder;
 
     let s1 = SomeEvent {
         id: 9,
@@ -54,13 +57,13 @@ fn main() {
             id: [0u8; 32].into(),
             da_height: 1,
             transactions_count: 1,
-            output_messages_count: 1,
+            message_receipt_count: 1,
             transactions_root: [0u8; 32].into(),
-            output_messages_root: [0u8; 32].into(),
             height: 1,
             prev_root: [0u8; 32].into(),
             time: 1,
             application_hash: [0u8; 32].into(),
+            message_receipt_root: [0u8; 32].into(),
         },
         transactions: vec![fuel::TransactionData {
             id: [0u8; 32].into(),
@@ -82,7 +85,7 @@ fn main() {
                     ptr: 2342143,
                     len: bytes1.len() as u64,
                     digest: [0u8; 32].into(),
-                    data: bytes1,
+                    data: Some(bytes1),
                     pc: 0,
                     is: 0,
                 },
@@ -102,7 +105,7 @@ fn main() {
                     ptr: 2342143,
                     len: bytes2.len() as u64,
                     digest: [0u8; 32].into(),
-                    data: bytes2,
+                    data: Some(bytes2),
                     pc: 0,
                     is: 0,
                 },
