+++ packages/fuel-indexer-tests/trybuild/simple_wasm.yaml
@@ -1,7 +1,7 @@
 namespace: test_namespace
 identifier: simple_wasm_executor
 abi: ./../fuel-indexer-tests/contracts/simple-wasm/out/debug/contracts-abi.json
-graphql_schema: ./../fuel-indexer-tests/components/indexers/simple-wasm/schema/simple_wasm.graphql
+graphql_schema: ./../fuel-indexer-tests/indexers/simple-wasm/schema/simple_wasm.graphql
 contract_id: ~
 module:
   wasm: ./../../../target/wasm32-unknown-unknown/release/simple_wasm.wasm
